FIX シンボリックリンクの作成に失敗する問題の修正

+++ app/startup.sh
@@ -3,7 +3,10 @@
 set -eu
 
 cd /usr/local/apache2/htdocs/
-ln -s /volumes/music_files music_files
+if [ ! -e "music_files" ]; then
+    ln -s /volumes/music_files music_files
+fi
+
 python3 -B /usr/src/app/feed_generator.py
 python3 -B /usr/src/app/scheduler.py &
 httpd -D FOREGROUND
